safety-extraction/requirements.txt

groq>=0.9.0
PyMuPDF>=1.24.0
sentence-transformers>=2.2.0
spacy>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database + Web viewer
psycopg2-binary>=2.9.0
pgvector>=0.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0

# python -m spacy download en_core_web_sm


